tripletex/requirements.txt
fastapi
uvicorn[standard]
requests
langgraph
langchain-google-genai
langchain-core
python-dotenv
